+++ app/controllers/OtherIncomeController.scala
@@ -62,11 +62,11 @@ class OtherIncomeController @Inject()(val baseConfig: BaseControllerConfig,
                 keystoreService.fetchIncomeSource() map {
                   case Some(incomeSource) => incomeSource.source match {
                     case IncomeSourceForm.option_business =>
-                      Redirect(controllers.business.routes.CurrentFinancialPeriodPriorController.show())
+                      Redirect(controllers.business.routes.BusinessAccountingPeriodController.show())
                     case IncomeSourceForm.option_property =>
                       Redirect(controllers.routes.TermsController.showTerms())
                     case IncomeSourceForm.option_both =>
-                      Redirect(controllers.business.routes.CurrentFinancialPeriodPriorController.show())
+                      Redirect(controllers.business.routes.BusinessAccountingPeriodController.show())
                   }
                 }
             }
