+++ app/views/business/accounting_period_date.scala.html
@@ -1,0 +1,71 @@
+@*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import play.api.data.Form
+@import models._
+@import forms.AccountingPeriodDateForm._
+@import helpers._
+@import uk.gov.hmrc.play.views.html.helpers.form
+@import config.AppConfig
+@import models.enums.{AccountingPeriodViewType, CurrentAccountingPeriodView}
+
+@(accountingPeriodForm: Form[AccountingPeriodModel], postAction: Call, backUrl: String, viewType: AccountingPeriodViewType, isEditMode: Boolean)(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
+
+@scriptElem = {
+    <script type="text/javascript" src="@routes.Assets.at("javascripts/jquery.min.js")"></script>
+    <script type="text/javascript" src="@routes.Assets.at("javascripts/jquery-ui.min.js")"></script>
+    <script type="text/javascript" src="@routes.Assets.at("javascripts/date-helper.js")"></script>
+}
+
+@views.html.templates.main_template(title = Messages("accounting_period.title"), bodyClasses = None, scriptElem = None) {
+    @back_link(backUrl)
+
+    @summaryErrorHelper(accountingPeriodForm)
+
+    <h1 class="heading-xlarge">
+        @if(viewType == CurrentAccountingPeriodView) {
+            @Messages("accounting_period.heading.current")
+        } else {
+            @Messages("accounting_period.heading.next")
+        }
+    </h1>
+
+    <p>@Messages("accounting_period.line_1")</p>
+
+    @form(action = postAction) {
+
+        @dateHelper(
+            field = accountingPeriodForm(startDate),
+            label = Messages("base.startDate"),
+            parentForm = accountingPeriodForm,
+            hintText = Messages("base.exampleDate")
+        )
+
+        @dateHelper(
+            field = accountingPeriodForm(endDate),
+            label = Messages("base.endDate"),
+            parentForm = accountingPeriodForm,
+            hintText = Messages("base.exampleDate")
+        )
+
+        @if(isEditMode) {
+            @continueButton(Messages("base.update"))
+        } else {
+            @continueButton()
+        }
+
+    }
+}
