+++ conf/app.routes
@@ -1,5 +1,5 @@
 # Assets
-GET         /assets/*file              controllers.Assets.at(path="/public", file)
+GET         /assets/*file                     controllers.Assets.at(path="/public", file)
 
 # Home controller
 GET         /                                 controllers.HomeController.index
