+++ conf/messages
@@ -38,11 +38,12 @@ property.income.line_1                                          = This is the am
 
 
 # Current Financial Period Prior page
-business.current_financial_period_prior.title                   = Did your current financial period start before 1 April 2017?
-business.current_financial_period_prior.heading                 = Did your current financial period start before 1 April 2017?
-business.current_financial_period_prior.line_1                  = This is the period that your latest income and expense records cover (also called your accounting period). Your financial period normally starts on 6 April, unless you’ve chosen to use your own dates.
-business.current_financial_period_prior.line_2                  = If you''re a new business, this is the date you started trading.
-business.current_financial_period_prior.hint                    = 1 April 2017 is when the Register for digital Income Tax updates services started.
+business.current_financial_period_prior.title                   = Did your current accounting period start before 1 April 2017?
+business.current_financial_period_prior.heading                 = Did your current accounting period start before 1 April 2017?
+business.current_financial_period_prior.line_1                  = 1 April 2017 is the date this service started.
+business.current_financial_period_prior.accordion               = What is an accounting period?
+business.current_financial_period_prior.accordion.line1         = This is the period that your latest income and expense records cover for Self Assessment.
+business.current_financial_period_prior.accordion.line2         = Your start date is usually 6 April (the same as the tax year), unless you selected a different date when you registered for Self Assessment.
 business.current_financial_period_prior.yes                     = Yes
 business.current_financial_period_prior.no                      = No
