+++ test/controllers/OtherIncomeErrorControllerSpec.scala
@@ -69,13 +69,13 @@ class OtherIncomeErrorControllerSpec extends ControllerBaseSpec with MockKeystor
       verifyKeystore(fetchIncomeSource = 1)
     }
 
-    s"redirect to '${controllers.business.routes.CurrentFinancialPeriodPriorController.show().url}' on the business journey" in {
+    s"redirect to '${controllers.business.routes.BusinessAccountingPeriodController.show().url}' on the business journey" in {
 
       setupMockKeystore(fetchIncomeSource = TestModels.testIncomeSourceBusiness)
 
       val goodRequest = callSubmit
 
-      redirectLocation(goodRequest) mustBe Some(controllers.business.routes.CurrentFinancialPeriodPriorController.show().url)
+      redirectLocation(goodRequest) mustBe Some(controllers.business.routes.BusinessAccountingPeriodController.show().url)
 
       await(goodRequest)
       verifyKeystore(fetchIncomeSource = 1)
@@ -93,13 +93,13 @@ class OtherIncomeErrorControllerSpec extends ControllerBaseSpec with MockKeystor
       verifyKeystore(fetchIncomeSource = 1)
     }
 
-    s"redirect to '${controllers.business.routes.CurrentFinancialPeriodPriorController.show().url}' on the both journey" in {
+    s"redirect to '${controllers.business.routes.BusinessAccountingPeriodController.show().url}' on the both journey" in {
 
       setupMockKeystore(fetchIncomeSource = TestModels.testIncomeSourceBoth)
 
       val goodRequest = callSubmit
 
-      redirectLocation(goodRequest) mustBe Some(controllers.business.routes.CurrentFinancialPeriodPriorController.show().url)
+      redirectLocation(goodRequest) mustBe Some(controllers.business.routes.BusinessAccountingPeriodController.show().url)
 
       await(goodRequest)
       verifyKeystore(fetchIncomeSource = 1)
