+++ test/controllers/business/BusinessNameControllerSpec.scala
@@ -120,8 +120,8 @@ class BusinessNameControllerSpec extends ControllerBaseSpec
   }
 
   "The back url" should {
-    s"point to ${controllers.business.routes.BusinessAccountingPeriodController.showAccountingPeriod().url}" in {
-      TestBusinessNameController.backUrl mustBe controllers.business.routes.BusinessAccountingPeriodController.showAccountingPeriod().url
+    s"point to ${controllers.business.routes.BusinessAccountingPeriodDateController.showAccountingPeriod().url}" in {
+      TestBusinessNameController.backUrl mustBe controllers.business.routes.BusinessAccountingPeriodDateController.showAccountingPeriod().url
     }
   }
