+++ test/utils/TestModels.scala
@@ -80,9 +80,9 @@ object TestModels extends Implicits {
 
   lazy val testIncomeSourceBoth = IncomeSourceModel(IncomeSourceForm.option_both)
 
-  lazy val testIsCurrentPeriod = CurrentFinancialPeriodPriorModel(CurrentFinancialPeriodPriorForm.option_no)
+  lazy val testIsCurrentPeriod = AccoutingPeriodPriorModel(AccountingPeriodPriorForm.option_no)
 
-  lazy val testIsNextPeriod = CurrentFinancialPeriodPriorModel(CurrentFinancialPeriodPriorForm.option_yes)
+  lazy val testIsNextPeriod = AccoutingPeriodPriorModel(AccountingPeriodPriorForm.option_yes)
 
   lazy val testOtherIncomeNo = OtherIncomeModel(OtherIncomeForm.option_no)
